Fix bug in writeBuffer: Ensure that each page write doesn't overrun the page boundary.
If the initial address to write to isn't on a page boundary, than the first chunk to write
must be smaller to only write to the end of that page.

Minor additional fix: toWrite was a uint16_t, but all calculations in this module are done
uint32_t, and the writeMemory() expects uint32_t as well.