qspi/SAMD: Remove redundant CSMODE=NORELOAD
The CTRLB register assignment includes both CSMODE=NORELOAD and
CSMODE=LASTXFER, which is a conflict.

Because NORELOAD=0x0 and LASTXFER=0x1, ORing these produces the same
result as just LASTXFER, and so I assume that was what was intended here.
This should be functionally equivalent to the previous code, but perhaps
a little easier for future humans to read.